Write out parsed names to a text file

Write out parsed names to a text file

ts_write_names(df, path)

Arguments

df

Dataframe with parsed names

path

Path to write dataframe

Writes out parsed names in a format that can be used by taxon-tools (each part of the scientific name is separated by the pipe symbol (|), with one name per line).

Value

Path to parsed names

Examples

if (ts_tt_installed()) {
  parsed_names <- ts_parse_names(
    "Foogenus x barspecies var. foosubsp (L.) F. Bar")
  temp_file <- tempfile()
  ts_write_names(parsed_names, temp_file)
  readLines(temp_file)
  file.remove(temp_file)
}
#> [1] TRUE