Resolve synonyms in taxonomic names

After matching taxonomic names to a reference, some may match synonyms. This function resolves synonyms to their accepted names.

ts_resolve_names(query, ref_taxonomy)

Arguments

query
Character vector or dataframe; taxonomic names to be resolved. If a character vector, missing values not allowed and all values must be unique. If a dataframe, should be taxonomic names matched with ts_match_names().

ref_taxonomy
Dataframe; reference taxonomic data adhering to the Darwin Core standard with the following columns:

taxonID: Unique identifier for each taxon.

acceptedNameUsageID: If the taxon is a synonym, the unique identifier for the accepted name

taxonomicStatus: The status of the use of the scientificName as a label for the taxon. Allowed values include accepted name, ambiguous synonym, provisionally accepted name, synonym.

scientificName: The full scientific name of the taxon, with authorship and date information if known.

Value

Dataframe; results of resolving synonyms in matched taxonomic names. Includes the following columns:

query: Query name

accepted_name: Accepted name after resolving synonyms

match_type: Type of match (for a summary of match types, see taxon-tools manual)

status: Taxonomic status of the matched reference name (same as taxonomicStatus in ref_taxonomy)

reference: Matched reference name

Names that could not be matched or resolve to multiple, different synonyms have NA for accepted_name.

Details

query can take as input either a character vector of taxonomic names, or the output of ts_match_names(). If the former, it will run ts_match_names() to match the query to ref_taxonomy, then resolve synonyms. If the latter, the scientific names in ref_taxonomy should be the same used as reference with ts_match_names() (this is not checked).

ref_taxonomy must be taxonomic data adhering to the Darwin Core standard. Darwin Core includes many terms, but only four (taxonID, acceptedNameUsageID, taxonomicStatus, and scientificName) are required for this function.

Examples

# Load reference taxonomy in Darwin Core format
data(filmy_taxonomy)

ts_resolve_names("Gonocormus minutum", filmy_taxonomy)

#> Error in rethrow_call(c_processx_exec, command, c(command, args), pty,     pty_options, connections, env, windows_verbatim_args, windows_hide_window,     windows_detached_process, private, cleanup, wd, encoding,     paste0("PROCESSX_", private$tree_id, "=YES")): cannot start processx process 'parsenames' (system error 2, No such file or directory) @unix/processx.c:608 (processx_exec)